feat: add /continue-as-guest GET route
This route sets the req.session.isGuest to true, and then redirects to
/ route.

This will let personalize the view for the / route, to have an initial
index page where the user can decide whether to continue as a guest, log
in or sign up. The choice is persisted until the session is active.

When the user log in, this req.session.isGuest is set to false.