feat: add guest/user icon on the header
The icon is personalized based on whether the current client is a guest,
simple registered user, member or admin.

If it is a registered user (including member or admin), a link to the
user profile is inserted, too.

In the profile page, there is a button to logout, hence that button is
removed from the header, for clarity.