feat: add deleteMessage POST route
A 404 error page is shown in case of unauthorized attempt to delete a
message, ie, by someone which is not the author or an admin. This is
done instead of 401, 403, 409 custom errors, to obfuscate this url.
Indeed, the corresponding url cannot be bookmarked.

The actual update of the database is not implemented, yet.