feat!: ENFORCE PRG PATTERN on error responses
When the server detects an error which is thrown, it responds with
the custom error view already implemented.

The probem is that when the request generating the error is not a GET
request, the PRG pattern is not enforced.

Eg, if such error is generated in a POST /login request, the client
receives a response to that route. Refreshing the page triggers another
POST request being sent.

Although this is "harmless" in most views (eg, the already logged in 409
error is thrown both for the GET and POST route), it might be annoying
for the user to confirm that the module must be sent back again.

So, in the error handling middleware in app.js, the METHOD of the
request is first checked: if it is not GET, a redirectTo URL is
obtained, error data are saved as flash messages on notGetError field,
and a redirect request is sent.

The redirectToUrl is obtained this way:
1 it is equal to res.locals.onNotGetErrorRedirectTo if it is defined.
  The validity / safety of this url must be already checked.
2 else, use a fallback route (/error)

It is assumed that 1 has ways to report the error messages.

For /error route, if "notGetError" error is present, the error
middleware is called with such error. Else, a 404 error is thrown.

Note: the definition of res.locals.onNotGetErrorRedirectTo is not
implemented, yet.