feat: add variable to keep track of each user last read message
This will be used to notified the user / guest whether there are any
new message (unread since last visit or page navigation).

For users, this info in stored in the users database as in the field
"last_read_message_id" of each user.

Instead, for guests, this is stored in the session, in the field
"lastReadMessageId", as it will last only until the guest is connected.