feat: set up WebSocket connection for real-time message list updates
This connection is designed to notify connected clients about new
messages sent to the server, not to receive messages from clients.

Currently, the server broadcasts a simple text to all connected clients
on a periodic interval, to trigger page refresh. A future update will
optimize this by broadcasting only when the message list has changed,
and by sending only the portion of the HTML message list with the new
messages, to be prepended to the message list.

On the client side:
- The WebSocket connection is established on page load.
- The connection is suspended when page visibility becomes hidden to
  save resources.
- Incoming HTML replaces the message list dynamically, avoiding full
  page reloads (TODO).

Note: since the page is just refreshed, the functionality of unread
message count is broken (ie, a new message becomes read as soon as the
next refresh is issued). This will be fixed later on.

This is based on the implementation of this functionality in:
https://github.com/frarosset/mini-message-board