fix: avoid updating lastReadMessageId when the refresh is triggered by other users db updates
Whenever the WebSocket sends a message to notify the clients that the
db has been updated, the clients set a wsRefresh cookie and reloads the
current page.

On the server, if a wsRefresh cookie is present, it means the request is
triggered by websocket, so just set a locals wsRefresh to true, so that
a temporary lastReadMessageId (saved in the session) is updated instead
of the actual one for the user. Otherwise, just set the
lastReadMessageId of the user as the above-mentioned temporary one, if
present.

The same is done when the client reopens a page toggling its visibility
state back to visible, which triggers a refresh to look for any new
message.

This way the user can see all the new messages, which are set as 'read'
only when the user directly navigates to the page.

Note: instead of sending the updated html of the list, a full reload
is used. Otherwise, authentication is needed for the websocket
connection, which is out of the scope of this project.