Implement user tabs refresh sync via BroadcastChannel
Refresh other tabs of the same user whenever it loads a new page that
gets updated (unless a refresh is triggered by websockets).

Eg, when there are 3 tabs opens, showing 5 unread messages in each. If
the user refreshed one of these, the unread messages become 0 for that
tab, and the other tabs are refreshed and show the correct updated
info.

The following operations are done:
- Assign unique socketId per tab using sessionStorage to distinguish
  between tabs and persist it through reloads.
- On page load, broadcast through BroadcastChannel 'refresh-other-tabs'
  message to sync sibling tabs (unless the broadcast has not been
  suspended: see next point)
- Suppress infinite refresh loops using "suppressBroadcast" data stored
  on sessionStorage, which is either true or false
- Tabs refreshing from a wss-redirect do not trigger such broadcast,
  because it gets suspended